requirements.txt

scikit-learn 
pandas
scikit-surprise
